+++ plugins/inputs/execd/shim/goshim.go
@@ -24,10 +24,8 @@ import (
 type empty struct{}
 
 var (
-	stdout        io.Writer = os.Stdout
-	stdin         io.Reader = os.Stdin
-	forever                 = 100 * 365 * 24 * time.Hour
-	envVarEscaper           = strings.NewReplacer(
+	forever       = 100 * 365 * 24 * time.Hour
+	envVarEscaper = strings.NewReplacer(
 		`"`, `\"`,
 		`\`, `\\`,
 	)
@@ -45,11 +43,19 @@ type Shim struct {
 	Inputs            []telegraf.Input
 	gatherPromptChans []chan empty
 	metricCh          chan telegraf.Metric
+
+	stdin  io.Reader
+	stdout io.Writer
+	stderr io.Writer
 }
 
 // New creates a new shim interface
 func New() *Shim {
-	return &Shim{}
+	return &Shim{
+		stdin:  os.Stdin,
+		stdout: os.Stdout,
+		stderr: os.Stderr,
+	}
 }
 
 // AddInput adds the input to the shim. Later calls to Run() will run this input.
@@ -108,7 +114,7 @@ func (s *Shim) Run(pollInterval time.Duration) error {
 		s.gatherPromptChans = append(s.gatherPromptChans, gatherPromptCh)
 		wg.Add(1) // one per input
 		go func(input telegraf.Input) {
-			startGathering(ctx, input, acc, gatherPromptCh, pollInterval)
+			s.startGathering(ctx, input, acc, gatherPromptCh, pollInterval)
 			if serviceInput, ok := input.(telegraf.ServiceInput); ok {
 				serviceInput.Stop()
 			}
@@ -141,7 +147,7 @@ loop:
 				return fmt.Errorf("failed to serialize metric: %s", err)
 			}
 			// Write this to stdout
-			fmt.Fprint(stdout, string(b))
+			fmt.Fprint(s.stdout, string(b))
 		}
 	}
 
@@ -163,7 +169,7 @@ func (s *Shim) stdinCollectMetricsPrompt(ctx context.Context, cancel context.Can
 		close(collectMetricsPrompt)
 	}()
 
-	scanner := bufio.NewScanner(stdin)
+	scanner := bufio.NewScanner(s.stdin)
 	// for every line read from stdin, make sure we're not supposed to quit,
 	// then push a message on to the collectMetricsPrompt
 	for scanner.Scan() {
@@ -201,7 +207,7 @@ func (s *Shim) collectMetrics(ctx context.Context) {
 	}
 }
 
-func startGathering(ctx context.Context, input telegraf.Input, acc telegraf.Accumulator, gatherPromptCh <-chan empty, pollInterval time.Duration) {
+func (s *Shim) startGathering(ctx context.Context, input telegraf.Input, acc telegraf.Accumulator, gatherPromptCh <-chan empty, pollInterval time.Duration) {
 	if pollInterval == PollIntervalDisabled {
 		return // don't poll
 	}
@@ -218,11 +224,11 @@ func startGathering(ctx context.Context, input telegraf.Input, acc telegraf.Accu
 			return
 		case <-gatherPromptCh:
 			if err := input.Gather(acc); err != nil {
-				fmt.Fprintf(os.Stderr, "failed to gather metrics: %s", err)
+				fmt.Fprintf(s.stderr, "failed to gather metrics: %s", err)
 			}
 		case <-t.C:
 			if err := input.Gather(acc); err != nil {
-				fmt.Fprintf(os.Stderr, "failed to gather metrics: %s", err)
+				fmt.Fprintf(s.stderr, "failed to gather metrics: %s", err)
 			}
 		}
 	}
@@ -269,12 +275,7 @@ func LoadConfig(filePath *string) ([]telegraf.Input, error) {
 		return nil, err
 	}
 
-	loadedInputs, err := loadConfigIntoInputs(md, conf.Inputs)
-
-	if len(md.Undecoded()) > 0 {
-		fmt.Fprintf(stdout, "Some plugins were loaded but not used: %q\n", md.Undecoded())
-	}
-	return loadedInputs, err
+	return loadConfigIntoInputs(md, conf.Inputs)
 }
 
 func expandEnvVars(contents []byte) string {